+++ noodles-vcf/examples/vcf_query.rs
@@ -18,7 +18,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let raw_region = args.next().expect("missing region");
 
     let mut reader = vcf::indexed_reader::Builder::default().build_from_path(src)?;
-    let header = reader.read_header()?.parse()?;
+    let header = reader.read_header()?;
 
     let region = raw_region.parse()?;
     let query = reader.query(&header, &region)?;