+++ noodles-vcf/CHANGELOG.md
@@ -14,6 +14,11 @@
 
 ### Changed
 
+  * vcf/reader: Change `Reader::read_header` to return a parsed header
+    (`Header`).
+
+    This no longer returns a raw string.
+
   * vcf/reader: Change `Reader::read_record` to read into a `Record` buffer.
 
     This now requires a `Header` and `Record` buffer. It no longer returns the